+++ models/firewall.py
@@ -1,10 +1,11 @@
 from utils import Iptables, Page
+from .db import BasicModel
 
 iptables = Iptables()
 
-class Firewall:
-    def __init__(self):
-        pass 
+class Firewall(BasicModel):
+    def __init__(self, model='firewall_policies'):
+        super().__init__(model=model)
 
     def get(self, page=1):
         data_list = iptables.get_rules()
@@ -17,8 +18,12 @@ class Firewall:
         ip_class = request_data['ip_class']
         protocol = request_data['protocol']
         port = request_data['port']
-        block = request_data['block']
-        iptables.post_rule(ip, ip_class=ip_class, protocol=protocol, port=port, block=block)
+        block = request_data['block'] 
+        firewall_policy = self.collection.find_one(request_data)
+        if not firewall_policy:
+            iptables.post_rule(ip, ip_class=ip_class, protocol=protocol, port=port, block=block)
+            self.collection.insert_one(request_data)
+            self.db['tickets'].update_many({'ip': ip}, {'$set': {'fix': '차단', 'fix_timestamp': request_data['timestamp']}})
 
     def delete(self, request_data={}):
         ip = request_data['ip']
@@ -26,4 +31,8 @@ class Firewall:
         protocol = request_data['protocol']
         port = request_data['port']
         block = request_data['block']
+        firewall_policy = self.collection.find_one(request_data)
         iptables.delete_rule(ip, ip_class=ip_class, protocol=protocol, port=port, block=block)
+        if firewall_policy:
+            self.collection.delete_one(request_data)
+            self.db['tickets'].update_many({'ip': ip}, {'$set': {'fix': '해제', 'fix_timestamp': ''}})